% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_nth.R
\name{every_nth}
\alias{every_nth}
\title{Helper function for ggplot to set axis labels to every nth}
\usage{
every_nth(n = x)
}
\arguments{
\item{every_nth(n}{= x) x is number of skips in x or y axis values}
}
\value{
Axis labels in a ggplot object
}
\description{
Helper function for ggplot to set axis labels to every nth
}
\examples{
scale_x_discrete(breaks = every_nth(n = 3))
}
