% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Finds observations not in a vector, using the negate function for \%in\%}
\usage{
%notin% c(" ", " "))
}
\arguments{
\item{c("", }{"") Items in defined vector that the function will filter out}
}
\value{
A new column or filtered data-frame based on the \%notin\% selection
}
\description{
Finds observations not in a vector, using the negate function for \%in\%
}
\examples{
filter(state \%notin\% c("PA", "MD"))
}
