% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctchange.R
\name{pctchange}
\alias{pctchange}
\title{Calculates s percent changes}
\usage{
pctchange(x)
}
\arguments{
\item{x}{uses lag function to calculate change from 1 row prior. Assumes long data.}
}
\value{
Calculation of percent change from prior row, either in console
or as new column if used with mutate()
}
\description{
Calculates s percent changes
}
\examples{
r_and_d_main \%>\%
  filter(sector == "Total funding") \%>\%
  group_by(field) \%>\%
  mutate(pct_chg_funding = pctchange(dkk_million)) \%>\%
  ungroup()
}
